refactor(capture): split the warnings module and stop it importing v1
The AI endpoints need their own warning conversion, and dropping it into
one flat file alongside legacy and v1 would make a 900-line module with
four pipelines interleaved. Splitting first keeps that PR pure addition.

The split is between two jobs, not by file size:

- Context projections turn one pipeline's request type into a
  WarningRequestContext. They are inseparable from the types they read,
  so they live with their pipeline.
- Emit helpers and type mappers decide which warning a failure becomes
  and which details ride along. All four pipelines' details land in one
  customer-facing table whose reader cannot tell which pipeline served a
  request, so the payloads must not drift. Sibling files keep drift
  visible in review.

That resolves an inverted dependency: ingestion_warnings.rs imported
crate::v1::context::RequestContext, so the shared module reached into
the isolated v1 subtree. v1_request_context was the only reason, and its
only callers were inside v1, so it becomes RequestContext::warning_context
next to the sdk_lib_and_version it projects. No ingestion_warnings/v1.rs:
a named file there would promote the leak into a declared structure.

Also drops now-redundant name prefixes: legacy_request_context becomes
legacy::request_context.

No behavior change.
