+++ rust/capture/src/events/analytics.rs
@@ -25,8 +25,10 @@ use crate::{
     events::overflow_stamping::stamp_overflow_reason,
     global_rate_limiter::{GlobalRateLimitKey, GlobalRateLimiter},
     ingestion_warnings::{
-        emit_distinct_id_truncated_warning, emit_processing_abort_warning, emit_rate_limit_warning,
-        legacy_request_context,
+        emit_rate_limit_warning,
+        legacy::{
+            emit_distinct_id_truncated_warning, emit_processing_abort_warning, request_context,
+        },
     },
     prometheus::{report_clock_skew, report_dropped_events},
     router, sinks,
@@ -532,7 +534,7 @@ async fn process_events_inner(
             if warned_event_count > 0 {
                 emit_rate_limit_warning(
                     ingestion_warning_emitter.as_deref(),
-                    &legacy_request_context(context),
+                    &request_context(context),
                     CAPTURE_LEGACY_RATE_LIMIT,
                     &warned_distinct_ids,
                     warned_event_count,
@@ -590,7 +592,7 @@ async fn process_events_inner(
     if truncated_count > 0 {
         emit_distinct_id_truncated_warning(
             ingestion_warning_emitter.as_deref(),
-            &legacy_request_context(context),
+            &request_context(context),
             truncated_sample.filter(|_| truncated_count == 1),
             truncated_count,
         );
